DishCategoryDomainServiceTest moved to the dish package
